Parenthesised takes pStmts

+++ src/test/scala/flux/Printer.scala
@@ -4,7 +4,7 @@ class PrinterTest extends munit.FunSuite {
   test("Parenthesised single") {
     val printed = Printer().print(
       Parenthesised(
-        Single("stmt"), begin = Some("("), end = Some(")"), sep = Some(",")
+        Single("stmt"), begin = Some(Single("(")), end = Some(Single(")")), sep = Some(Single(","))
       )
     )
     assertEquals(printed, Print("(stmt)"))
@@ -12,7 +12,7 @@ class PrinterTest extends munit.FunSuite {
   test("Parenthesised many") {
     val printed = Printer().print(
       Parenthesised(
-        Many(List(Single("stmt0"), Single("stmt1"))), begin = Some("("), end = Some(")"), sep = Some(",")
+        Many(List(Single("stmt0"), Single("stmt1"))), begin = Some(Single("(")), end = Some(Single(")")), sep = Some(Single(","))
       )
     )
     assertEquals(printed, Print("(stmt0,stmt1)"))
